refactor(personaplex): isolate validation components
Move the stateful Mimi encoder and codec weight translation out of the family entry point. Pin the lazy Full-Duplex-Bench scoring environment and keep the paired metric implementation independent of SciPy while preserving its golden outputs.

The existing FP16 L0 bundle contract still builds the streaming encoder in the required FP32 precision; no CI case or validation threshold is changed.
