+++ python/tensorrt_model_connect/families/personaplex/full_duplex_evaluator_verify.py
@@ -5,19 +5,25 @@ from importlib.metadata import version
 
 import nemo
 import nemo.collections.asr as nemo_asr
+import numpy
 import scipy
 import silero_vad
 import soundfile
 import torch
 import torchaudio
 
+assert version("huggingface-hub") == "1.22.0"
 assert version("nemo_toolkit") == "2.7.3"
-assert version("scipy") == "1.18.0"
+assert version("numpy") == "1.26.4"
+assert version("scipy") == "1.15.3"
 assert version("silero-vad") == "6.2.1"
 assert version("soundfile") == "0.14.0"
+assert version("torch") == "2.12.0+cu130"
+assert version("torchaudio") == "2.11.0+cu130"
 assert nemo is not None
 assert nemo_asr is not None
-assert scipy.__version__ == "1.18.0"
+assert numpy.__version__ == "1.26.4"
+assert scipy.__version__ == "1.15.3"
 assert silero_vad.__version__ == "6.2.1"
 assert soundfile.__version__ == "0.14.0"
 assert torch.cuda.is_available()
@@ -25,10 +31,14 @@ assert torchaudio is not None
 print(
     " ".join(
         (
+            f"huggingface-hub={version('huggingface-hub')}",
             f"nemo_toolkit={version('nemo_toolkit')}",
+            f"numpy={numpy.__version__}",
             f"scipy={scipy.__version__}",
             f"silero-vad={silero_vad.__version__}",
             f"soundfile={soundfile.__version__}",
+            f"torch={torch.__version__}",
+            f"torchaudio={version('torchaudio')}",
         )
     )
 )
