+++ python/tensorrt_model_connect/families/personaplex/python_profile_requirements/full_duplex_evaluator.lock.txt
@@ -1,4 +1,6 @@
+huggingface-hub==1.22.0
 nemo_toolkit[asr]==2.7.3
-scipy==1.18.0
+numpy==1.26.4
+scipy==1.15.3
 silero-vad==6.2.1
 soundfile==0.14.0
